+++ Plugin/Selva.GH/Features/FileIO/OBSOLETE/OBSOLETE_FileFromPath_UntilV0_11_0.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
 using Grasshopper.Kernel;
@@ -8,11 +7,14 @@ using Selva.GH.Features.FileIO.Services;
 using Selva.GH.Properties;
 using Selva.GH.Utilities;
 
-namespace Selva.GH.Features.FileIO.Components;
+namespace Selva.GH.Features.FileIO.OBSOLETE;
 
-public class GH_FileFromPath : GH_Component, ISelvaFileOutput
+/// <summary>
+///     Obsolete File From Path component (until v0.11.0). Replaced by the version with a Metadata input.
+/// </summary>
+public class OBSOLETE_FileFromPath_UntilV0_11_0 : GH_Component, ISelvaFileOutput
 {
-    public GH_FileFromPath()
+    public OBSOLETE_FileFromPath_UntilV0_11_0()
         : base("File From Path", "FilePath",
             "Reads a file from disk and outputs it as file data.",
             "Selva", "IO")
@@ -21,7 +23,9 @@ public class GH_FileFromPath : GH_Component, ISelvaFileOutput
 
     protected override Bitmap Icon => Resources.PathToFile;
 
-    public override Guid ComponentGuid => new Guid("A66F2DA2-800B-460D-9996-7C13BDEE4553");
+    public override Guid ComponentGuid => new Guid("F2B8D4A6-C3E7-4B1F-9D5A-8E2C6F4A1B3D");
+
+    public override GH_Exposure Exposure => GH_Exposure.hidden;
 
     public override void CreateAttributes()
     {
@@ -35,10 +39,6 @@ public class GH_FileFromPath : GH_Component, ISelvaFileOutput
             GH_ParamAccess.item, "");
         pManager.AddTextParameter("Sub Folder", "Folder", "Optional subfolder path for storage", GH_ParamAccess.item,
             "");
-        pManager.AddTextParameter("Metadata", "M",
-            "Optional metadata as \"key=value\" lines (e.g. author=felix). Rides along with the file for downstream tagging/indexing.",
-            GH_ParamAccess.list);
-        pManager[3].Optional = true;
     }
 
     protected override void RegisterOutputParams(GH_OutputParamManager pManager)
@@ -59,9 +59,6 @@ public class GH_FileFromPath : GH_Component, ISelvaFileOutput
 
         DA.GetData(1, ref nameOverride);
         DA.GetData(2, ref subFolder);
-
-        var metadataLines = new List<string>();
-        DA.GetDataList(3, metadataLines);
 
         if (!File.Exists(path))
         {
@@ -91,8 +88,7 @@ public class GH_FileFromPath : GH_Component, ISelvaFileOutput
             Data = Convert.ToBase64String(bytes),
             FileType = extension,
             IsBase64Encoded = true,
-            SubFolder = subFolder ?? "",
-            Metadata = FileMetadataParser.Parse(metadataLines)
+            SubFolder = subFolder ?? ""
         };
 
         DA.SetData(0, new FileDataGoo(fileData));